Move Distributor ordering logic to Firm
unintegrated for now
